Add clickable links to Travis CI branches in task cards.
Now it is possible to open the Travis CI test results by clicking on the
badge in the trello card, same as for the Appveyor tests.

Unfortunately I have not been able to find a good way of showing just
the desired branch, so now we show the branch overview instead. It lists
the most recent branches at the top which should be good enough.